Included some editor dominos to pageflow-support gem
