--- spec/support/dominos/editor/entry_outline.rb
@@ -1,11 +1,0 @@
-module Dom
-  module Editor
-    class EntryOutline < Domino
-      selector 'sidebar .chapters.outline'
-
-      def self.await!
-        find!
-      end
-    end
-  end
-end
