+++ spec/support/dominos/editor/public_dominos.rb
@@ -1,0 +1,7 @@
+require 'pageflow/dom'
+
+module Dom
+  module Editor
+    include Pageflow::Dom::Editor
+  end
+end
