+++ spec/support/pageflow/dom.rb
@@ -1,6 +1,7 @@
 require 'domino'
 
 require 'pageflow/dom/admin'
+require 'pageflow/dom/editor'
 
 # Dominos to be used by plugin tests.
 module Pageflow
