+++ spec/support/pageflow/dom/editor.rb
@@ -1,0 +1,12 @@
+require 'pageflow/dom/editor/entry_outline'
+require 'pageflow/dom/editor/sidebar'
+require 'pageflow/dom/editor/file_import_chooser'
+require 'pageflow/dom/editor/file_import_dialog'
+
+module Pageflow
+  module Dom
+    # Dominos for editor ui elements and helper methods.
+    module Editor
+    end
+  end
+end
