+++ spec/support/pageflow/dom/editor/entry_outline.rb
@@ -1,0 +1,13 @@
+module Pageflow
+  module Dom
+    module Editor
+      class EntryOutline < Domino
+        selector 'sidebar .chapters.outline'
+
+        def self.await!
+          find!
+        end
+      end
+    end
+  end
+end
