+++ spec/support/pageflow/dom/editor/public_dominos.rb
@@ -1,0 +1,7 @@
+require 'pageflow/dom'
+
+module Dom
+  module Editor
+    include Dom::Editor
+  end
+end
